fix(mcp-local): clear the Clippy errors the pipewire failure was hiding
With libpipewire installed the job gets far enough to actually lint
tauri-plugin-mcp, and 31 errors surface. They are not new code: Rust 1.97
extended `collapsible_if` to suggest let-chains for edition-2024 crates,
and `-D warnings` makes every one fatal. Nothing noticed because the
build died in libspa-sys's build script before Clippy reached this crate.

29 are mechanical and come straight from `clippy --fix`: 21 collapsible_if,
3 question_mark, plus useless_format, unnecessary_lazy_evaluations,
unnecessary_cast, and map_clone.

Two needed a decision:

- `IpcBuffer::push` trips too_many_arguments at 8/7. Its sibling
  `push_with_origin`, which it forwards all eight fields to, already
  carries the same allow — so this matches the call already made rather
  than restructuring one of a delegating pair.
- `command_stats` guarded a division with `if dur_count > 0`. Replaced by
  `checked_div`, which is equivalent here: the field is Option<u64>,
  starts None from Default, and is assigned nowhere else, so the None the
  divisor-zero branch now writes is the None the guard used to leave.

Note `cargo fmt --check --manifest-path apps/desktop/Cargo.toml` covers
only the root package, so this crate is not rustfmt-clean — running
`cargo fmt -p tauri-plugin-mcp` reformats 32 files. Left alone here; the
`clippy --fix` output satisfies the fmt check as it stands.

Verified with the toolchain CI resolves to (1.97.1, local stable was five
months behind): both `cargo fmt --check` and `cargo clippy -- -D warnings`
on apps/desktop/Cargo.toml exit 0.

+++ apps/desktop/tauri-plugin-mcp-local/src/platform/macos.rs
@@ -51,7 +51,7 @@ fn get_all_windows_cg() -> Vec<WindowInfo> {
         let array: CFArray = CFArray::wrap_under_create_rule(window_list);
 
         for i in 0..array.len() {
-            let dict_ref = match array.get(i as isize) {
+            let dict_ref = match array.get(i) {
                 Some(r) => r,
                 None => continue,
             };
